moved factorio credentials to its own file && tested all mods routes
